requirements.txt
apache-airflow == 2.2.1
